+++ Changelog.md
@@ -1,3 +1,7 @@
+## Qlibs 0.4.2
+ - Fix: not every window can now be closed.
+ - Fix: App skips rendering if window width or height is equal to zero. (Caused crash on windows)
+
 ## Qlibs 0.4.-1
  - Font system being reworked.
  - Fixed weird word wrapping behavior.